Create windscape.cpp
In this project, we aim at building a scenario, as called windscape, which will giving us the working animated knowledge of how the traffic lights work, in according with OpenGL and we have built a scene where the windmills , sunrays and all the day to day things are animated for a treat to our viewers. This project is basically built by firstly assembling all the components of the scenario, like the car, the windmills, sun, sunrays, car wheels, mountains, traffic light, zebra crossing, etc. and putting them together to build this animation.
Firstly, our window appears, in which we have introduced ourselves and then our code is
designed in such a way that it takes input from the keyboard to return the main page - the
windscape to the user.
When the main page arrives, the car moves because, initially our light is green and
windmills rotate in the clockwise direction and the sun in the anticlockwise direction.
If we want to stop the car from surpassing the zebra crossing, we need to click our mouse so
that we set the value of traffic light to red, after which it stops, however , the movement of
windmills and the sun is unaffected by the input given by the mouse .
This project tested our basic knowledge on OpenGL.